feat(codex): implement model discovery and probing functionality
- Added `probeModels` function to discover Codex models from a custom endpoint or fallback models.
- Introduced tests for `probeModels` to validate model discovery and error handling.
- Updated `model-connection-registry-core` to export `probeModels` and `modelsUrl`.
- Refactored existing tests to accommodate new model discovery logic.
